Add the ability to switch off the creation of symlink, depending on the RPM provider

+++ common/torque2/client/config.tpl
@@ -301,34 +301,35 @@ include { 'components/altlogrotate/config' };
 include { 'components/symlink/config' };
 
 "/software/components/symlink/links" = { 
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qstat",
-                               "target", "/usr/bin/qstat-torque",
-                               "replace", nlist("all","yes"),                                                                                                                                                                                                                                                                                                                                                                            
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qsub",
-                               "target", "/usr/bin/qsub-torque",
-                               "replace", nlist("all","yes"),
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qhold",
-                               "target", "/usr/bin/qhold-torque",
-                               "replace", nlist("all","yes"),
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qrls",
-                               "target", "/usr/bin/qrls-torque",
-                               "replace", nlist("all","yes"),
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qalter",
-                               "target", "/usr/bin/qalter-torque",
-                               "replace", nlist("all","yes"),
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qselect",
-                               "target", "/usr/bin/qselect-torque",
-                               "replace", nlist("all","yes"),
-                              );  
-  SELF[length(SELF)] =   nlist("name", "/usr/bin/qdel",
-                               "target", "/usr/bin/qdel-torque",
-                               "replace", nlist("all","yes"),
-                              );  
+  if (TORQUE_PROVIDER=='lal') {
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qstat",
+                                 "target", "/usr/bin/qstat-torque",
+                                 "replace", nlist("all","yes"),                                                 );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qsub",
+                                 "target", "/usr/bin/qsub-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qhold",
+                                 "target", "/usr/bin/qhold-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qrls",
+                                 "target", "/usr/bin/qrls-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qalter",
+                                 "target", "/usr/bin/qalter-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qselect",
+                                 "target", "/usr/bin/qselect-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+    SELF[length(SELF)] =   nlist("name", "/usr/bin/qdel",
+                                 "target", "/usr/bin/qdel-torque",
+                                 "replace", nlist("all","yes"),
+                                );
+  };
   SELF;
 };
